docs(README.md): update repository name and description to reflect that it is a NodeJS Clean Architecture Template and add instructions to remove original remote and add new remote origin
The repository name and description have been updated to reflect that this is a NodeJS Clean Architecture Template. Additionally, instructions have been added to remove the original remote and add a new remote origin to the cloned repository. This is to ensure that the cloned repository is not linked to the original repository and that changes made to the cloned repository are pushed to the correct repository.